feat: attach login functionality to login page
