+++ webapp/src/loginData.ts
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+const isLogIn = writable(false);
+
+export default isLogIn;
