--- webapp/src/user.js
@@ -1,5 +1,0 @@
-import { writable } from 'svelte/store';
-
-const user = writable(null);
-
-export default user;
