+++ webapp/tailwind.config.ts
@@ -1,11 +1,11 @@
-import type { Config } from "tailwindcss";
+import type { Config } from 'tailwindcss';
 
 export default {
-    content: ["./src/**/*.{html,js,svelte,ts}"],
+    content: ['./src/**/*.{html,js,svelte,ts}'],
 
     theme: {
-        extend: {},
+        extend: {}
     },
 
-    plugins: [require("@tailwindcss/typography")],
+    plugins: []
 } as Config;
